+++ docs/whats_new.rst
@@ -29,6 +29,8 @@ Enhancements
   attributes (`attrs`). Together with `cf_xarray <https://cf-xarray.readthedocs.io/en/latest/flags.html>`__
   these can be leveraged to select single (``mask.cf == "CNA"``) or multiple (``mask.cf.isin``)
   regions (:pull:`361`, :issue:`346`).
+- Added :py:func:`flatten_3D_mask` - a helper function to flatten 3D masks to 2D masks
+  (:issue:`399`).
 - The masking functions (e.g. :py:meth:`Regions.mask`) now warn if the `units` of the
   coordinates(``lat.attrs["units"]`` ) are given as "radians" (:issue:`279`).
 - Better error when passing a single region without wrapping it into a list or tuple (:issue:`372`).